Remove uchar typedef
typedef creates just an alias and does not increase type safety.

+++ buffer-iter.c
@@ -1,12 +1,13 @@
 #include "buffer.h"
+#include "uchar.h"
 
-unsigned int buffer_get_char(struct block_iter *bi, uchar *up)
+unsigned int buffer_get_char(struct block_iter *bi, unsigned int *up)
 {
 	struct block_iter tmp = *bi;
 	return buffer_next_char(&tmp, up);
 }
 
-unsigned int buffer_next_char(struct block_iter *bi, uchar *up)
+unsigned int buffer_next_char(struct block_iter *bi, unsigned int *up)
 {
 	unsigned int offset = bi->offset;
 
@@ -28,7 +29,7 @@ unsigned int buffer_next_char(struct block_iter *bi, uchar *up)
 	return bi->offset - offset;
 }
 
-unsigned int buffer_prev_char(struct block_iter *bi, uchar *up)
+unsigned int buffer_prev_char(struct block_iter *bi, unsigned int *up)
 {
 	unsigned int offset = bi->offset;
